Update test 02 to check multiple types, remove 03
