--- test/xrt/03_mul_shim_i16/run.lit
@@ -1,6 +1,0 @@
-// (c) Copyright 2024 Advanced Micro Devices, Inc.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-// REQUIRES: ryzen_ai, valid_xchess_license
-// RUN: %python %S/gen.py
-// RUN: %run_on_ipu %python %S/run.py aie.xclbin
